requirements.txt
python-dotenv==1.0.1
pytest==8.3.3
pytest-html-reporter==0.2.9
pytest-metadata==3.1.1
requests==2.32.3